move resource to constant

+++ features/transaction/add_on_gifting/src/main/java/com/tokopedia/addongifting/data/getaddonsavedstate/GetAddOnSavedStateResponse.kt
@@ -22,7 +22,7 @@ data class AddOnDataResponse(
         @SerializedName("add_on_qty")
         val addOnQty: Int = 0,
         @SerializedName("add_on_price")
-        val addOnPrice: Double = 0.0,
+        val addOnPrice: Long = 0,
         @SerializedName("add_on_metadata")
         val addOnMetadata: AddOnMetadataResponse = AddOnMetadataResponse()
 )
